Deleted old and irrelevant charts, separated out scripts by reports, continued cleaning code
